refactor(skills): split ClawHub lifecycle
